refactor: extract browser resolution policy and transport
